MainPlayerFragment.kt and MainPlayerViewModel.kt moved to presentation package
